// src/app/first-case-shared/calculation-types.ts
export class CalculationData {
  u: number;
  r1: number;
  c1: number;
  r2: number;
  c2: number;
  armLength: number;
  armSurface: number;
  upperBodyLength: number;
  upperBodySurface: number;
}

export class CalculationResult {
  I: number;
  armEnergyDensity: number;
  bodyEnergyDensity: number;
}